requirements.txt
certifi
charset-normalizer
click
colorama
filelock
fsspec
huggingface-hub
idna
Jinja2
joblib
MarkupSafe
mpmath
networkx
nltk
numpy
opencv-python
packaging
pdf2image
pillow
PyArabic
PyQt5
PyQt5-Qt5
PyQt5_sip
pytesseract
PyYAML
regex
requests
safetensors
scikit-learn
scipy
six
sympy
threadpoolctl
tokenizers
torch
torchvision
tqdm
transformers
typing_extensions
urllib3
Whoosh

